% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_boxplots.R
\name{normalization_boxplots}
\alias{normalization_boxplots}
\title{normalization_boxplots}
\usage{
normalization_boxplots(e2, f2, p2, color1 = "treatment",
  color2 = "mx class", sampleName = "phenotypeindex")
}
\description{
stat
}
\examples{
normalization_boxplots()
}

