% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_dealwithmissing.R
\name{normalization_dealwithmissing}
\alias{normalization_dealwithmissing}
\title{normalization_dealwithmissing}
\usage{
normalization_dealwithmissing(e, f, p, missindex = NA, removemiss = TRUE,
  miss_rate, tolmissingperc = 50, missingreplacemethod = "half minimum",
  missgroup = "organ,treatment")
}
\description{
stat
}
\examples{
normalization_dealwithmissing()
}

