% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_downloadnormalization.R
\name{normalization_downloadnormalization}
\alias{normalization_downloadnormalization}
\title{normalization_downloadnormalization}
\usage{
normalization_downloadnormalization(e1, f1, p1, samplewisenorm = "mTIC",
  samplewisenormindex = "knownornotknown", transformation = "log",
  loga = 1, logbase = 2, power = 1/2, scaling = "autoscaling")
}
\description{
stat
}
\examples{
normalization_downloadnormalization()
}

