% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_gg_color_hue.R
\name{normalization_gg_color_hue}
\alias{normalization_gg_color_hue}
\title{normalization_gg_color_hue}
\usage{
normalization_gg_color_hue(n)
}
\arguments{
\item{n}{number of colors.}
}
\description{
normalization_gg_color_hue
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

