% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_normalization.R
\name{normalization_normalization}
\alias{normalization_normalization}
\title{normalization_normalization}
\usage{
normalization_normalization(e1, f1, p1, samplewisenorm = "mTIC",
  samplewisenormindex = "knownornotknown", transformation = "log",
  loga = 1, logbase = 2, power = 1/2, scaling = "autoscaling")
}
\description{
stat
}
\examples{
normalization_normalization()
}

