% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_numofmissing.R
\name{normalization_numofmissing}
\alias{normalization_numofmissing}
\title{normalization_numofmissing}
\usage{
normalization_numofmissing(e, f, p, missindex = NA,
  compoundName = "Binbase name", tolmissingperc = 50)
}
\description{
stat
}
\examples{
normalization_numofmissing()
}

