% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_uploadfile.R
\name{normalization_uploadfile}
\alias{normalization_uploadfile}
\title{normalization_uploadfile}
\usage{
normalization_uploadfile(file)
}
\description{
stat
}
\examples{
normalization_uploadfile()
}

