% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_onefactormultigroupspower.R
\name{samplesize_onefactormultigroupspower}
\alias{samplesize_onefactormultigroupspower}
\title{samplesize_onefactormultigroupspower}
\usage{
samplesize_onefactormultigroupspower(k = 3, effectsize = 0.8,
  sig_level = 0.05, power = 0.8, forplot = FALSE, samplerange = NULL,
  effectsizerange = NULL)
}
\description{
stat
}
\examples{
samplesize_onefactormultigroupspower()
}

