% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_onefactormultipairedgroupspower.R
\name{samplesize_onefactormultipairedgroupspower}
\alias{samplesize_onefactormultipairedgroupspower}
\title{samplesize_onefactormultipairedgroupspower}
\usage{
samplesize_onefactormultipairedgroupspower(m = 3, effectsize = 0.8,
  sig_level = 0.05, power = 0.8, corr = 0.5, epsilon = 0.75,
  forplot = FALSE, samplerange = NULL, effectsizerange = NULL)
}
\description{
stat
}
\examples{
samplesize_onefactormultipairedgroupspower()
}

