% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_population_power_plot.R
\name{samplesize_onefactortwogroupspower}
\alias{samplesize_onefactortwogroupspower}
\title{samplesize_onefactortwogroupspower}
\usage{
samplesize_onefactortwogroupspower(effectsize = 0.8, sig_level = 0.05,
  power = 0.8)
}
\description{
stat
}
\examples{
samplesize_onefactortwogroupspower()
}

