% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_onefactortwopairedgroupspower.R
\name{samplesize_onefactortwopairedgroupspower}
\alias{samplesize_onefactortwopairedgroupspower}
\title{samplesize_onefactortwopairedgroupspower}
\usage{
samplesize_onefactortwopairedgroupspower(effectsize = 0.8, sig_level = 0.05,
  power = 0.8)
}
\description{
stat
}
\examples{
samplesize_onefactortwopairedgroupspower()
}

