% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_twofactorindindpower.R
\name{samplesize_twofactorindindpower}
\alias{samplesize_twofactorindindpower}
\title{samplesize_twofactorindindpower}
\usage{
samplesize_twofactorindindpower(k1 = 3, k2 = 3, effectsize = 0.8,
  sig_level = 0.05, power = 0.8, forplot = FALSE, samplerange = NULL,
  effectsizerange = NULL)
}
\description{
stat
}
\examples{
samplesize_twofactorindindpower()
}

