% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_twofactorindpairedpower.R
\name{samplesize_twofactorindpairedpower}
\alias{samplesize_twofactorindpairedpower}
\title{samplesize_twofactorindpairedpower}
\usage{
samplesize_twofactorindpairedpower(k = 3, m = 3, effectsize = 0.8,
  sig_level = 0.05, power = 0.8, corr = 0.5, epsilon = 0.75,
  forplot = FALSE, samplerange = NULL, effectsizerange = NULL)
}
\description{
stat
}
\examples{
samplesize_twofactorindpairedpower()
}

