% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_friedman_test_with_post_hoc.R
\name{stat_friedman_test_with_post_hoc}
\alias{stat_friedman_test_with_post_hoc}
\title{stat_friedman_test_with_post_hoc}
\usage{
stat_friedman_test_with_post_hoc(formu, data, to.print.friedman = F,
  to.post.hoc.if.signif = T, to.plot.parallel = F, to.plot.boxplot = F,
  signif.P = 0.05, color.blocks.in.cor.plot = F, jitter.Y.in.cor.plot = F)
}
\description{
stat_friedman_test_with_post_hoc
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

