% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_population_power_plot.R
\name{two_population_power_plot}
\alias{two_population_power_plot}
\title{two_population_power_plot}
\usage{
two_population_power_plot(effectsize = 0.8, n = seq(2, 100, by = 2),
  sig_level = 0.05)
}
\description{
stat
}
\examples{
two_population_power_plot()
}

