% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_boxplots.R
\name{univariateanalysis_boxplots}
\alias{univariateanalysis_boxplots}
\title{univariateanalysis_boxplots}
\usage{
univariateanalysis_boxplots(e2, f2, p2, color1 = "treatment",
  color2 = "mx class", sampleName = "phenotypeindex")
}
\description{
stat
}
\examples{
univariateanalysis_boxplots()
}

