% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_dealwithmissing.R
\name{univariateanalysis_dealwithmissing}
\alias{univariateanalysis_dealwithmissing}
\title{univariateanalysis_dealwithmissing}
\usage{
univariateanalysis_dealwithmissing(e, f, p, missindex = NA,
  removemiss = TRUE, miss_rate, tolmissingperc = 50,
  missingreplacemethod = "half minimum", missgroup = "organ,treatment")
}
\description{
stat
}
\examples{
univariateanalysis_dealwithmissing()
}

