% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_gg_color_hue.R
\name{univariateanalysis_gg_color_hue}
\alias{univariateanalysis_gg_color_hue}
\title{univariateanalysis_gg_color_hue}
\usage{
univariateanalysis_gg_color_hue(n)
}
\arguments{
\item{n}{number of colors.}
}
\description{
univariateanalysis_gg_color_hue
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

