% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_multiIndependentGroups.R
\name{univariateanalysis_multiIndependentGroups}
\alias{univariateanalysis_multiIndependentGroups}
\title{univariateanalysis_multiIndependentGroups}
\usage{
univariateanalysis_multiIndependentGroups(e2, f2, p2, group1 = "treatment",
  para = T, paraposthoc = "games.howell", equalvariance = F,
  nonpara = T, nonpposthoc = "t test", nonpposthocadj = "bonferroni",
  multicore = TRUE)
}
\description{
stat
}
\examples{
univariateanalysis_multiIndependentGroups()
}

