% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_normalization.R
\name{univariateanalysis_normalization}
\alias{univariateanalysis_normalization}
\title{univariateanalysis_normalization}
\usage{
univariateanalysis_normalization(e1, f1, p1, samplewisenorm = "mTIC",
  samplewisenormindex = "knownornotknown", transformation = "log",
  loga = 1, logbase = 2, power = 1/2, scaling = "autoscaling")
}
\description{
stat
}
\examples{
univariateanalysis_normalization()
}

