% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_PCA.R
\name{univariateanalysis_PCA}
\alias{univariateanalysis_PCA}
\title{univariateanalysis_PCA}
\usage{
univariateanalysis_PCA(e2, f2, p2, color = "treatment",
  shape = "mx class id")
}
\description{
stat
}
\examples{
univariateanalysis_PCA()
}

