% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_summarizegroup.R
\name{univariateanalysis_summarizegroup}
\alias{univariateanalysis_summarizegroup}
\title{univariateanalysis_summarizegroup}
\usage{
univariateanalysis_summarizegroup(e2, f2, p2, group1 = "species",
  group2 = "treatment", test = "two independent groups")
}
\description{
stat
}
\examples{
univariateanalysis_summarizegroup()
}

