% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_twoIndependentGroups.R
\name{univariateanalysis_twoIndependentGroups}
\alias{univariateanalysis_twoIndependentGroups}
\title{univariateanalysis_twoIndependentGroups}
\usage{
univariateanalysis_twoIndependentGroups(e2, f2, p2, group1 = "treatment",
  ttestpara = T, ttestparaadj = "fdr", ttestequalvariance = F,
  ttestnonpara = T, ttestnonparaadj = "fdr", multicore = TRUE)
}
\description{
stat
}
\examples{
univariateanalysis_twoIndependentGroups()
}

