% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_twoPairedGroups.R
\name{univariateanalysis_twoPairedGroups}
\alias{univariateanalysis_twoPairedGroups}
\title{univariateanalysis_twoPairedGroups}
\usage{
univariateanalysis_twoPairedGroups(e2, f2, p2, group1 = "phenotype",
  pairedttestpara = T, pairedttestparaadj = "fdr", pairedttestnonpara = T,
  pairedttestnonparaadj = "fdr", multicore = TRUE)
}
\description{
stat
}
\examples{
univariateanalysis_twoPairedGroups()
}

