% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_uploadfile.R
\name{univariateanalysis_uploadfile}
\alias{univariateanalysis_uploadfile}
\title{univariateanalysis_uploadfile}
\usage{
univariateanalysis_uploadfile(file)
}
\description{
stat
}
\examples{
univariateanalysis_uploadfile()
}

