% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_PCA.R
\name{normalization_PCA}
\alias{normalization_PCA}
\title{normalization_PCA}
\usage{
normalization_PCA(e2, f2, p2, color = "treatment", shape = "mx class id")
}
\description{
stat
}
\examples{
normalization_PCA()
}

