% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_visullizeeffectsize.R
\name{samplesize_visullizeeffectsize}
\alias{samplesize_visullizeeffectsize}
\title{samplesize_visullizeeffectsize}
\usage{
samplesize_visullizeeffectsize(means = c(-1, 0, 1), sds = c(1, 2, 1),
  pooled_variance = NULL, sigma_mu = NULL, effectsizevalue = NULL)
}
\description{
stat
}
\examples{
samplesize_visullizeeffectsize()
}

