% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateanalysis_numofmissing.R
\name{univariateanalysis_numofmissing}
\alias{univariateanalysis_numofmissing}
\title{univariateanalysis_numofmissing}
\usage{
univariateanalysis_numofmissing(e, f, p, missindex = NA)
}
\description{
stat
}
\examples{
univariateanalysis_numofmissing()
}

